Allow table row count to be greater than Integer.MAX_VALUE (2^31-1) by refactoring numRows from int to long

+++ src/main/java/net/sourceforge/schemaspy/model/Database.java
@@ -310,7 +310,7 @@ public class Database {
         final String type;
         final String remarks;
         final String viewSql;
-        final int numRows;  // -1 if not determined
+        final long numRows;  // -1 if not determined
 
         /**
          * @param schema
@@ -320,7 +320,7 @@ public class Database {
          * @param text optional textual SQL used to create the view
          * @param numRows number of rows, or -1 if not determined
          */
-        BasicTableMeta(String schema, String name, String type, String remarks, String text, int numRows)
+        BasicTableMeta(String schema, String name, String type, String remarks, String text, long numRows)
         {
             this.schema = schema;
             this.name = name;
@@ -365,7 +365,7 @@ public class Database {
                     String remarks = getOptionalString(rs, clazz + "_comment");
                     String text = forTables ? null : getOptionalString(rs, "view_definition");
                     String rows = forTables ? getOptionalString(rs, "table_rows") : null;
-                    int numRows = rows == null ? -1 : Integer.parseInt(rows);
+                    long numRows = rows == null ? -1 : Long.parseLong(rows);
 
                     basics.add(new BasicTableMeta(sch, name, clazz, remarks, text, numRows));
                 }
